Feistel modular(for implementation)
Will encode/decode character string of length 1 to 1024(presumably)
given that the size argument is an odd number.  Originally had
difficulties with terminating zeros corrupting data, but as long as size
is odd, all is well.  If you look at the keyset generation, it's very
rudimentary but provides well-randomized, reversable results.